rename and clarify serialization specs (our serialization of columns vs activemodel record serialization)
